+++ Assets/SuperNamespaace/AssetType1.cs
@@ -3,7 +3,7 @@ using UnityEngine;
 namespace SuperNamespaace
 {
     [CreateAssetMenu(menuName = "Examples/Type1")]
-    public class AssetType1 : ScriptableObject
+    public class AssetType1 : ScriptableObject, IAsset
     {
         
     }