+++ Assets/SuperNamespaace/AssetType2.cs
@@ -3,7 +3,7 @@ using UnityEngine;
 namespace SuperNamespaace
 {
     [CreateAssetMenu(menuName = "Examples/Type2")]
-    public class AssetType2 : ScriptableObject
+    public class AssetType2 : ScriptableObject, IAsset
     {
         
     }