+++ Assets/SuperNamespaace/IAsset.cs
@@ -1,0 +1,7 @@
+namespace SuperNamespaace
+{
+    public interface IAsset
+    {
+        
+    }
+}